+++ src/components/BottomNavigation/testmap.tsx
@@ -5,6 +5,9 @@ import {
   faUser,
 } from "@fortawesome/free-solid-svg-icons";
 
+// lib
+import { APP_ROUTES } from "lib";
+
 // types
 import type { BottomNavigationItemType } from "./types";
 
@@ -12,28 +15,28 @@ export const bottomMap: BottomNavigationItemType[] = [
   {
     id: "home",
     label: "Home",
-    to: "/",
+    to: APP_ROUTES.ROOT,
     icon: faHome,
     position: "left",
   },
   {
     id: "notes",
     label: "Notes",
-    to: "/notes",
+    to: APP_ROUTES.NOTES,
     icon: faFile,
     position: "left",
   },
   {
     id: "categories",
     label: "Categories",
-    to: "/categories",
+    to: APP_ROUTES.CATEGORIES,
     icon: faTags,
     position: "right",
   },
   {
     id: "profile",
     label: "Profile",
-    to: "/profile",
+    to: APP_ROUTES.PROFILE,
     icon: faUser,
     position: "right",
   },
