+++ src/lib/index.ts
@@ -12,5 +12,8 @@ export * from "./api";
 // utils
 export * from "./utils";
 
+// routes
+export * from "./routes";
+
 // errors
 export * from "./errors";
