+++ src/providers/Auth/index.ts
@@ -1,3 +1,5 @@
 export * from "./types";
 export * from "./AuthProvider";
 export * from "./authContext";
+export * from "./useAuthContext";
+export * from "./useOptionalAuthContext";
