+++ src/providers/BottomNavAction/index.ts
@@ -1,7 +1,5 @@
 export { BottomNavActionProvider } from "./BottomNavActionProvider";
-export {
-  useBottomNavAction,
-  useOptionalBottomNavAction,
-  useRegisterBottomNavAction,
-} from "./useBottomNavAction";
-export type { BottomNavActionRegistrationType } from "./useBottomNavAction";
+export { useBottomNavAction } from "./useBottomNavAction";
+export { useOptionalBottomNavAction } from "./useOptionalBottomNavAction";
+export { useRegisterBottomNavAction } from "./useRegisterBottomNavAction";
+export type { BottomNavActionRegistrationType } from "./types";
