+++ src/providers/BottomNavAction/types.ts
@@ -1,0 +1,6 @@
+import { BottomNavigationCenterActionType } from "../../components";
+
+export type BottomNavActionRegistrationType =
+  | BottomNavigationCenterActionType
+  | (() => void)
+  | null;
